refactor(conf): split validate.go into focused domain files
Split the 2,017-line validate.go into five focused files by domain:

- validate.go (158 lines): shared types (ValidationError,
  ValidationResult), constants, regex patterns, ValidateSettings
  orchestrator
- validate_audio.go (553 lines): stream, audio source, RTSP,
  quiet hours, export bitrate, normalization validation
- validate_services.go (710 lines): BirdNET, Birdweather, MQTT,
  telemetry, web server, webhook, notification validation
- validate_security.go (225 lines): security, TLS mode, OIDC,
  AutoTLS hostname validation
- validate_realtime.go (414 lines): realtime settings, sound level,
  retention, dashboard, weather, species tracking validation

Pure file-organization change with zero behavior change. Follows
the same pattern as PR #2867 (config.go split).

Completes Forgejo #492 (conf package refactor).
