+++ internal/conf/validate.go
@@ -4,19 +4,9 @@ package conf
 
 import (
 	"fmt"
-	"net"
-	"net/url"
-	"os/exec"
 	"regexp"
-	"slices"
-	"strconv"
-	"strings"
-	"text/template"
-	"time"
 
 	"github.com/tphakala/birdnet-go/internal/errors"
-	"github.com/tphakala/birdnet-go/internal/logger"
-	"github.com/tphakala/birdnet-go/internal/templatefuncs"
 )
 
 // MinSoundLevelInterval is the minimum sound level interval in seconds to prevent excessive CPU usage
@@ -62,34 +52,6 @@ var (
 	gpsCoordPattern = regexp.MustCompile(`^:?[+-]?\d+\.\d+\s*,\s*[+-]?\d+(\.\d+)?$|^:?[+-]?\d+(\.\d+)?\s*,\s*[+-]?\d+\.\d+$`)
 )
 
-// Audio gain limits in dB
-const (
-	MinAudioGain = -40.0 // Minimum allowed audio gain in dB
-	MaxAudioGain = 40.0  // Maximum allowed audio gain in dB
-)
-
-// Playback gain limits in dB (UI preference, distinct from export gain)
-const (
-	MinPlaybackGain     = 0.0  // Minimum default playback gain
-	MaxPlaybackGain     = 24.0 // Maximum default playback gain (matches AudioSettingsButton GAIN_MAX_DB)
-	DefaultPlaybackGain = 0.0  // Safe default if invalid value provided
-)
-
-// Audio export type constants
-const (
-	AudioExportTypeWAV  = "wav"  // Lossless PCM audio
-	AudioExportTypeFLAC = "flac" // Lossless compressed audio
-	AudioExportTypeMP3  = "mp3"  // Lossy compressed audio
-	AudioExportTypeAAC  = "aac"  // Lossy compressed audio
-	AudioExportTypeOPUS = "opus" // Lossy compressed audio
-)
-
-// Stream validation constants
-const (
-	MaxStreamNameLength      = 64
-	MaxAudioSourceNameLength = 100
-)
-
 // ValidAudioModels contains recognized AI model identifiers.
 // Empty string is also valid (defaults to birdnet).
 var ValidAudioModels = map[string]bool{
@@ -98,55 +60,6 @@ var ValidAudioModels = map[string]bool{
 	ModelIDPerchV2: true,
 	ModelIDBat:     true,
 }
-
-// Quiet hours validation constants
-const (
-	MaxQuietHoursOffset = 180  // Maximum offset in minutes from sun event
-	MinQuietHoursOffset = -180 // Minimum offset in minutes from sun event
-)
-
-// Quiet hours mode constants
-const (
-	QuietHoursModeFixed = "fixed" // Fixed clock-based quiet hours
-	QuietHoursModeSolar = "solar" // Solar event-based quiet hours
-)
-
-// Solar event constants
-const (
-	SolarEventSunrise = "sunrise" // Sunrise solar event
-	SolarEventSunset  = "sunset"  // Sunset solar event
-)
-
-// ValidQuietHoursModes contains valid quiet hours mode values
-var ValidQuietHoursModes = map[string]bool{
-	QuietHoursModeFixed: true,
-	QuietHoursModeSolar: true,
-}
-
-// ValidSolarEvents contains valid solar event names
-var ValidSolarEvents = map[string]bool{
-	SolarEventSunrise: true,
-	SolarEventSunset:  true,
-}
-
-// ValidStreamTypes contains all supported stream types
-var ValidStreamTypes = map[string]bool{
-	StreamTypeRTSP: true,
-	StreamTypeHTTP: true,
-	StreamTypeHLS:  true,
-	StreamTypeRTMP: true,
-	StreamTypeUDP:  true,
-}
-
-// EBU R128 normalization limits
-const (
-	MinTargetLUFS    = -40.0 // Minimum target loudness in LUFS
-	MaxTargetLUFS    = -10.0 // Maximum target loudness in LUFS
-	MinLoudnessRange = 0.0   // Minimum loudness range in LU
-	MaxLoudnessRange = 20.0  // Maximum loudness range in LU
-	MinTruePeak      = -10.0 // Minimum true peak in dBTP
-	MaxTruePeak      = 0.0   // Maximum true peak in dBTP
-)
 
 // ValidationError represents a collection of validation errors
 type ValidationError struct {
@@ -176,607 +89,6 @@ type ValidationResult struct {
 	Errors     []string // Validation errors (fatal)
 	Warnings   []string // Non-fatal warnings
 	Normalized any      // Normalized/transformed config (type matches input)
-}
-
-// Validate validates a single stream configuration
-func (s *StreamConfig) Validate() error {
-	// Name is required
-	name := strings.TrimSpace(s.Name)
-	if name == "" {
-		return fmt.Errorf("stream name is required")
-	}
-
-	// Name length limit (check trimmed name for consistency)
-	if len(name) > MaxStreamNameLength {
-		return fmt.Errorf("stream name '%s' exceeds maximum length of %d characters", name, MaxStreamNameLength)
-	}
-
-	// URL is required
-	if strings.TrimSpace(s.URL) == "" {
-		return fmt.Errorf("stream URL is required for '%s'", s.Name)
-	}
-
-	// Validate stream type
-	if !ValidStreamTypes[s.Type] {
-		return fmt.Errorf("invalid stream type '%s' for '%s': must be one of rtsp, http, hls, rtmp, udp", s.Type, s.Name)
-	}
-
-	// Validate transport (only tcp/udp allowed, empty defaults to tcp)
-	if s.Transport != "" && s.Transport != "tcp" && s.Transport != "udp" {
-		return fmt.Errorf("invalid transport '%s' for '%s': must be tcp or udp", s.Transport, s.Name)
-	}
-
-	// Validate URL scheme matches type
-	if err := s.validateURLScheme(); err != nil {
-		return err
-	}
-
-	// Validate quiet hours if enabled
-	if err := ValidateQuietHours(&s.QuietHours, fmt.Sprintf("stream '%s'", s.Name)); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ValidateQuietHours validates a quiet hours configuration
-func ValidateQuietHours(qh *QuietHoursConfig, context string) error {
-	if qh == nil || !qh.Enabled {
-		return nil
-	}
-
-	// Validate mode
-	if !ValidQuietHoursModes[qh.Mode] {
-		return fmt.Errorf("%s: quiet hours mode must be 'fixed' or 'solar', got '%s'", context, qh.Mode)
-	}
-
-	switch qh.Mode {
-	case QuietHoursModeFixed:
-		// Validate start time format
-		if _, err := time.Parse("15:04", qh.StartTime); err != nil {
-			return fmt.Errorf("%s: quiet hours start time must be in HH:MM format, got '%s'", context, qh.StartTime)
-		}
-		// Validate end time format
-		if _, err := time.Parse("15:04", qh.EndTime); err != nil {
-			return fmt.Errorf("%s: quiet hours end time must be in HH:MM format, got '%s'", context, qh.EndTime)
-		}
-
-	case QuietHoursModeSolar:
-		// Validate start event
-		if !ValidSolarEvents[qh.StartEvent] {
-			return fmt.Errorf("%s: quiet hours start event must be 'sunrise' or 'sunset', got '%s'", context, qh.StartEvent)
-		}
-		// Validate end event
-		if !ValidSolarEvents[qh.EndEvent] {
-			return fmt.Errorf("%s: quiet hours end event must be 'sunrise' or 'sunset', got '%s'", context, qh.EndEvent)
-		}
-		// Validate offsets
-		if qh.StartOffset < MinQuietHoursOffset || qh.StartOffset > MaxQuietHoursOffset {
-			return fmt.Errorf("%s: quiet hours start offset must be between %d and %d minutes, got %d", context, MinQuietHoursOffset, MaxQuietHoursOffset, qh.StartOffset)
-		}
-		if qh.EndOffset < MinQuietHoursOffset || qh.EndOffset > MaxQuietHoursOffset {
-			return fmt.Errorf("%s: quiet hours end offset must be between %d and %d minutes, got %d", context, MinQuietHoursOffset, MaxQuietHoursOffset, qh.EndOffset)
-		}
-	}
-
-	return nil
-}
-
-// validateURLScheme checks URL scheme matches declared stream type
-func (s *StreamConfig) validateURLScheme() error {
-	urlLower := strings.ToLower(s.URL)
-
-	switch s.Type {
-	case StreamTypeRTSP:
-		if !strings.HasPrefix(urlLower, "rtsp://") && !strings.HasPrefix(urlLower, "rtsps://") {
-			return fmt.Errorf("stream '%s': RTSP type requires rtsp:// or rtsps:// URL", s.Name)
-		}
-	case StreamTypeHTTP:
-		if !strings.HasPrefix(urlLower, "http://") && !strings.HasPrefix(urlLower, "https://") {
-			return fmt.Errorf("stream '%s': HTTP type requires http:// or https:// URL", s.Name)
-		}
-	case StreamTypeHLS:
-		if !strings.HasPrefix(urlLower, "http://") && !strings.HasPrefix(urlLower, "https://") {
-			return fmt.Errorf("stream '%s': HLS type requires http:// or https:// URL", s.Name)
-		}
-	case StreamTypeRTMP:
-		if !strings.HasPrefix(urlLower, "rtmp://") && !strings.HasPrefix(urlLower, "rtmps://") {
-			return fmt.Errorf("stream '%s': RTMP type requires rtmp:// or rtmps:// URL", s.Name)
-		}
-	case StreamTypeUDP:
-		if !strings.HasPrefix(urlLower, "udp://") && !strings.HasPrefix(urlLower, "rtp://") {
-			return fmt.Errorf("stream '%s': UDP type requires udp:// or rtp:// URL", s.Name)
-		}
-	}
-
-	return nil
-}
-
-// ApplyStreamDefaults sets default transport for RTSP/RTMP streams that have an empty
-// transport field. This handles the case where users write the new streams: YAML format
-// directly without specifying per-stream transport — the global RTSPSettings.Transport
-// (defaulting to "tcp") is propagated to each applicable stream.
-func (r *RTSPSettings) ApplyStreamDefaults() {
-	globalTransport := r.Transport
-	if globalTransport == "" {
-		globalTransport = DefaultTransport
-	}
-	for _, stream := range r.AllStreams() {
-		if stream.Transport == "" && (stream.Type == StreamTypeRTSP || stream.Type == StreamTypeRTMP) {
-			stream.Transport = globalTransport
-		}
-	}
-}
-
-// ValidateStreams validates the streams collection for uniqueness and individual validity
-func (r *RTSPSettings) ValidateStreams() error {
-	names := make(map[string]bool)
-	urls := make(map[string]bool)
-
-	for i, stream := range r.AllStreams() {
-		// Validate individual stream
-		if err := stream.Validate(); err != nil {
-			return fmt.Errorf("stream %d: %w", i+1, err)
-		}
-
-		// Check for duplicate names (case-insensitive)
-		nameLower := strings.ToLower(strings.TrimSpace(stream.Name))
-		if names[nameLower] {
-			return fmt.Errorf("duplicate stream name: '%s'", stream.Name)
-		}
-		names[nameLower] = true
-
-		// Check for duplicate URLs (trimmed for consistency)
-		urlTrimmed := strings.TrimSpace(stream.URL)
-		if urls[urlTrimmed] {
-			return fmt.Errorf("stream '%s' has a duplicate URL: '%s'", stream.Name, stream.URL)
-		}
-		urls[urlTrimmed] = true
-	}
-
-	return nil
-}
-
-// Validate validates a single audio source configuration.
-// It normalizes whitespace on Name, Device, and Model in-place.
-func (a *AudioSourceConfig) Validate() error {
-	// Normalize fields in-place so downstream code sees trimmed values.
-	a.Name = strings.TrimSpace(a.Name)
-	a.Device = strings.TrimSpace(a.Device)
-	a.Model = strings.TrimSpace(a.Model)
-
-	// Name is required
-	if a.Name == "" {
-		return fmt.Errorf("audio source name is required")
-	}
-	if len(a.Name) > MaxAudioSourceNameLength {
-		return fmt.Errorf("audio source name '%s' exceeds maximum length of %d characters", a.Name, MaxAudioSourceNameLength)
-	}
-
-	// Device is required
-	if a.Device == "" {
-		return fmt.Errorf("audio source device is required for '%s'", a.Name)
-	}
-
-	// Reject device strings that look like GPS coordinates. Users have
-	// pasted values like ":45.5,-120.5" into the device field (probably
-	// intended for the location/coordinates setting) which bypasses
-	// argument parsing in ALSA/pipewire/pulse and causes the audio
-	// engine to fail forever on every startup. Catch it up front with
-	// a clear error pointing at the right setting.
-	if gpsCoordPattern.MatchString(a.Device) {
-		return fmt.Errorf("audio source '%s': device %q looks like GPS coordinates; set latitude/longitude under birdnet.latitude and birdnet.longitude instead and pick a real audio device", a.Name, a.Device)
-	}
-
-	// Validate gain range
-	if a.Gain < MinAudioGain || a.Gain > MaxAudioGain {
-		return fmt.Errorf("audio source '%s': gain %.1f dB out of range [%.0f, +%.0f]", a.Name, a.Gain, MinAudioGain, MaxAudioGain)
-	}
-
-	// Validate model identifier
-	if !ValidAudioModels[a.Model] {
-		return fmt.Errorf("audio source '%s': unknown model '%s'", a.Name, a.Model)
-	}
-
-	// Validate per-source EQ if set
-	if a.Equalizer != nil {
-		for i, f := range a.Equalizer.Filters {
-			if f.Frequency <= 0 {
-				return fmt.Errorf("audio source '%s': equalizer filter %d has invalid frequency %.1f", a.Name, i+1, f.Frequency)
-			}
-		}
-	}
-
-	// Validate quiet hours
-	if err := ValidateQuietHours(&a.QuietHours, fmt.Sprintf("audio source '%s'", a.Name)); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ValidateSources validates all audio source configurations, including
-// duplicate name and device detection.
-func (a *AudioSettings) ValidateSources() error {
-	names := make(map[string]bool)
-	devices := make(map[string]bool)
-
-	for i := range a.Sources {
-		src := &a.Sources[i]
-		if err := src.Validate(); err != nil {
-			return fmt.Errorf("audio source %d: %w", i+1, err)
-		}
-
-		// Check for duplicate names (case-insensitive)
-		nameLower := strings.ToLower(strings.TrimSpace(src.Name))
-		if names[nameLower] {
-			return fmt.Errorf("duplicate audio source name: '%s'", src.Name)
-		}
-		names[nameLower] = true
-
-		// Check for duplicate devices (trimmed for consistency)
-		deviceTrimmed := strings.TrimSpace(src.Device)
-		if devices[deviceTrimmed] {
-			return fmt.Errorf("audio source '%s' has a duplicate device: '%s'", src.Name, src.Device)
-		}
-		devices[deviceTrimmed] = true
-	}
-
-	return nil
-}
-
-// ValidateBirdNETSettings performs BirdNET validation without side effects.
-// Returns normalized settings and any errors/warnings.
-// This pure function enables testing without log output or settings mutation.
-//
-// The private validateBirdNETSettings() calls this and handles side effects.
-func ValidateBirdNETSettings(cfg *BirdNETConfig) ValidationResult {
-	result := ValidationResult{Valid: true, Warnings: []string{}}
-	normalized := *cfg
-
-	// Sensitivity range check
-	if cfg.Sensitivity < 0 || cfg.Sensitivity > 1.5 {
-		result.Valid = false
-		result.Errors = append(result.Errors, "BirdNET sensitivity must be between 0 and 1.5")
-	}
-
-	// Threshold range check
-	if cfg.Threshold < 0 || cfg.Threshold > 1 {
-		result.Valid = false
-		result.Errors = append(result.Errors, "BirdNET threshold must be between 0 and 1")
-	}
-
-	// Overlap range check
-	if cfg.Overlap < 0 || cfg.Overlap > 2.99 {
-		result.Valid = false
-		result.Errors = append(result.Errors, "BirdNET overlap value must be between 0 and 2.99 seconds")
-	}
-
-	// Longitude range check
-	if cfg.Longitude < -180 || cfg.Longitude > 180 {
-		result.Valid = false
-		result.Errors = append(result.Errors, "BirdNET longitude must be between -180 and 180")
-	}
-
-	// Latitude range check
-	if cfg.Latitude < -90 || cfg.Latitude > 90 {
-		result.Valid = false
-		result.Errors = append(result.Errors, "BirdNET latitude must be between -90 and 90")
-	}
-
-	// Threads check
-	if cfg.Threads < 0 {
-		result.Valid = false
-		result.Errors = append(result.Errors, "BirdNET threads must be at least 0")
-	}
-
-	// RangeFilter model check - empty string, "latest", or "legacy" are valid
-	if cfg.RangeFilter.Model != "" && cfg.RangeFilter.Model != "latest" && cfg.RangeFilter.Model != "legacy" {
-		result.Valid = false
-		result.Errors = append(result.Errors, "RangeFilter model must be either empty (v2 default), 'latest', or 'legacy'")
-	}
-
-	// RangeFilter threshold check
-	if cfg.RangeFilter.Threshold < 0 || cfg.RangeFilter.Threshold > 1 {
-		result.Valid = false
-		result.Errors = append(result.Errors, "RangeFilter threshold must be between 0 and 1")
-	}
-
-	// Locale validation and normalization (pure transformation)
-	if cfg.Locale != "" {
-		normalizedLocale, err := NormalizeLocale(cfg.Locale)
-		if err != nil {
-			// Locale normalization fell back to default - this is a warning, not an error
-			message := fmt.Sprintf("BirdNET locale '%s' is not supported, will use fallback '%s'", cfg.Locale, normalizedLocale)
-			result.Warnings = append(result.Warnings, message)
-		}
-		// Update the normalized locale
-		normalized.Locale = normalizedLocale
-	}
-
-	result.Normalized = &normalized
-	return result
-}
-
-// ValidateBirdweatherSettings performs Birdweather validation without side effects.
-// Returns validation result with normalized settings.
-func ValidateBirdweatherSettings(settings *BirdweatherSettings) ValidationResult {
-	result := ValidationResult{Valid: true, Warnings: []string{}}
-	normalized := *settings
-
-	if settings.Enabled {
-		// Check if ID is provided when enabled
-		if settings.ID == "" {
-			result.Valid = false
-			result.Errors = append(result.Errors, "Birdweather ID is required when enabled")
-			// Suggest disabling
-			normalized.Enabled = false
-			result.Warnings = append(result.Warnings, "Birdweather will be disabled due to missing ID")
-		} else if !birdweatherIDPattern.MatchString(settings.ID) {
-			// Validate Birdweather ID format using precompiled regex
-			result.Valid = false
-			result.Errors = append(result.Errors, "Invalid Birdweather ID format: must be 24 alphanumeric characters")
-			normalized.Enabled = false
-			result.Warnings = append(result.Warnings, "Birdweather will be disabled due to invalid ID format")
-		}
-
-		// Check threshold range
-		if settings.Threshold < 0 || settings.Threshold > 1 {
-			result.Valid = false
-			result.Errors = append(result.Errors, "birdweather threshold must be between 0 and 1")
-		}
-
-		// Check location accuracy
-		if settings.LocationAccuracy < 0 {
-			result.Valid = false
-			result.Errors = append(result.Errors, "birdweather location accuracy must be non-negative")
-		}
-	}
-
-	result.Normalized = &normalized
-	return result
-}
-
-// ValidateWebhookProvider performs webhook provider validation without side effects.
-// Returns validation result with errors.
-func ValidateWebhookProvider(p *PushProviderConfig) ValidationResult {
-	result := ValidationResult{Valid: true}
-
-	if !p.Enabled {
-		result.Normalized = p
-		return result
-	}
-
-	// Webhook requires at least one endpoint
-	if len(p.Endpoints) == 0 {
-		result.Valid = false
-		result.Errors = append(result.Errors,
-			fmt.Sprintf("webhook provider '%s' requires at least one endpoint when enabled", p.Name))
-		result.Normalized = p
-		return result
-	}
-
-	// Validate custom template if specified.
-	// Use the shared templatefuncs.Funcs so Parse() accepts custom functions
-	// like "title" or "formatTime".
-	if p.Template != "" {
-		if _, err := template.New("validation").Funcs(templatefuncs.Funcs).Parse(p.Template); err != nil {
-			result.Valid = false
-			result.Errors = append(result.Errors,
-				fmt.Sprintf("webhook provider '%s': invalid template syntax: %v", p.Name, err))
-		}
-	}
-
-	// Validate each endpoint
-	for i := range p.Endpoints {
-		endpoint := &p.Endpoints[i]
-
-		// URL is required
-		if strings.TrimSpace(endpoint.URL) == "" {
-			result.Valid = false
-			result.Errors = append(result.Errors,
-				fmt.Sprintf("webhook provider '%s' endpoint %d: URL is required", p.Name, i))
-			continue
-		}
-
-		// URL must start with http:// or https://
-		if !strings.HasPrefix(endpoint.URL, "http://") && !strings.HasPrefix(endpoint.URL, "https://") {
-			result.Valid = false
-			result.Errors = append(result.Errors,
-				fmt.Sprintf("webhook provider '%s' endpoint %d: URL must start with http:// or https://", p.Name, i))
-		}
-
-		// Validate HTTP method if specified
-		if endpoint.Method != "" {
-			method := strings.ToUpper(endpoint.Method)
-			if method != "POST" && method != "PUT" && method != "PATCH" {
-				result.Valid = false
-				result.Errors = append(result.Errors,
-					fmt.Sprintf("webhook provider '%s' endpoint %d: method must be POST, PUT, or PATCH, got %s", p.Name, i, endpoint.Method))
-			}
-		}
-
-		// Validate timeout
-		if endpoint.Timeout < 0 {
-			result.Valid = false
-			result.Errors = append(result.Errors,
-				fmt.Sprintf("webhook provider '%s' endpoint %d: timeout must be non-negative", p.Name, i))
-		}
-	}
-
-	result.Normalized = p
-	return result
-}
-
-// ValidateMQTTSettings performs MQTT validation without side effects.
-// Returns validation result with errors.
-func ValidateMQTTSettings(settings *MQTTSettings) ValidationResult {
-	result := ValidationResult{Valid: true}
-
-	if !settings.Enabled {
-		result.Normalized = settings
-		return result
-	}
-
-	// Check if broker is provided when enabled
-	if settings.Broker == "" {
-		result.Valid = false
-		result.Errors = append(result.Errors, "MQTT broker URL is required when MQTT is enabled")
-	}
-
-	// Check if topic is provided when enabled
-	if settings.Topic == "" {
-		result.Valid = false
-		result.Errors = append(result.Errors, "MQTT topic is required when MQTT is enabled")
-	}
-
-	// Validate retry settings if enabled
-	if settings.RetrySettings.Enabled {
-		if settings.RetrySettings.MaxRetries < 0 {
-			result.Valid = false
-			result.Errors = append(result.Errors, "MQTT max retries must be non-negative")
-		}
-		if settings.RetrySettings.InitialDelay < 0 {
-			result.Valid = false
-			result.Errors = append(result.Errors, "MQTT initial delay must be non-negative")
-		}
-		if settings.RetrySettings.MaxDelay < settings.RetrySettings.InitialDelay {
-			result.Valid = false
-			result.Errors = append(result.Errors, "MQTT max delay must be greater than or equal to initial delay")
-		}
-		if settings.RetrySettings.BackoffMultiplier <= 0 {
-			result.Valid = false
-			result.Errors = append(result.Errors, "MQTT backoff multiplier must be positive")
-		}
-	}
-
-	result.Normalized = settings
-	return result
-}
-
-// ValidateWebServerSettings performs WebServer validation without side effects.
-// Returns validation result with errors.
-func ValidateWebServerSettings(settings *WebServerSettings) ValidationResult {
-	result := ValidationResult{Valid: true}
-
-	if settings.Enabled {
-		// Check if port is provided when enabled
-		if settings.Port == "" {
-			result.Valid = false
-			result.Errors = append(result.Errors, "WebServer port is required when enabled")
-		} else {
-			// Validate port is a valid number in range 1-65535
-			if port, err := strconv.Atoi(settings.Port); err != nil || port < 1 || port > 65535 {
-				result.Valid = false
-				result.Errors = append(result.Errors,
-					fmt.Sprintf("WebServer port must be a number between 1 and 65535, got %q", settings.Port))
-			}
-		}
-	}
-
-	// Validate BasePath (reverse proxy subpath prefix).
-	// Empty is allowed (disables basepath). Non-empty must be a safe URL-path prefix.
-	if err := validateBasePath(settings.BasePath); err != nil {
-		result.Valid = false
-		result.Errors = append(result.Errors, err.Error())
-	}
-
-	// Validate LiveStream settings
-	if settings.LiveStream.BitRate < 16 || settings.LiveStream.BitRate > 320 {
-		result.Valid = false
-		result.Errors = append(result.Errors,
-			fmt.Sprintf("LiveStream bitrate must be between 16 and 320 kbps, got %d", settings.LiveStream.BitRate))
-	}
-
-	if settings.LiveStream.SegmentLength < 1 || settings.LiveStream.SegmentLength > 30 {
-		result.Valid = false
-		result.Errors = append(result.Errors,
-			fmt.Sprintf("LiveStream segment length must be between 1 and 30 seconds, got %d", settings.LiveStream.SegmentLength))
-	}
-
-	result.Normalized = settings
-	return result
-}
-
-// validateBasePath verifies that a configured WebServer.BasePath is a safe URL-path prefix.
-// Empty is allowed and disables subpath routing. Non-empty values must start with "/",
-// contain only a restricted character set, and must not introduce path traversal,
-// protocol-relative URLs, backslashes, or scheme-like sequences.
-func validateBasePath(basePath string) error {
-	if basePath == "" {
-		return nil
-	}
-
-	if !strings.HasPrefix(basePath, "/") {
-		return fmt.Errorf("WebServer basepath must start with %q, got %q", "/", basePath)
-	}
-
-	// A single "/" is meaningless as a subpath prefix — require the caller to use empty instead.
-	if basePath == "/" {
-		return fmt.Errorf("WebServer basepath %q is meaningless; leave empty to disable subpath routing", basePath)
-	}
-
-	// Reject dangerous sequences outright. The sequence list aligns with the
-	// dangerous-pattern check in isValidBasePath (internal/api/v2/auth.go) so that
-	// YAML-configured and UI-supplied basepaths share the same rejection rules.
-	for _, bad := range []string{"..", "//", "\\", "://", "\n", "\r", "\x00"} {
-		if strings.Contains(basePath, bad) {
-			return fmt.Errorf("WebServer basepath contains disallowed sequence %q: %q", bad, basePath)
-		}
-	}
-
-	// Restrict to the same alphanumeric + /_- character set accepted elsewhere.
-	// This catches HTML/JS injection attempts, whitespace, and unicode surprises early.
-	for _, r := range basePath {
-		switch {
-		case r >= 'a' && r <= 'z',
-			r >= 'A' && r <= 'Z',
-			r >= '0' && r <= '9',
-			r == '/', r == '-', r == '_':
-			// allowed
-		default:
-			return fmt.Errorf("WebServer basepath contains disallowed character %q: %q", r, basePath)
-		}
-	}
-
-	return nil
-}
-
-// ValidateTelemetrySettings validates the telemetry endpoint configuration.
-// Returns validation result with errors if enabled and listen address is malformed.
-func ValidateTelemetrySettings(settings *TelemetrySettings) ValidationResult {
-	result := ValidationResult{Valid: true}
-
-	if !settings.Enabled {
-		return result
-	}
-
-	if settings.Listen == "" {
-		result.Valid = false
-		result.Errors = append(result.Errors, "telemetry listen address cannot be empty when enabled")
-		return result
-	}
-
-	_, portStr, err := net.SplitHostPort(settings.Listen)
-	if err != nil {
-		result.Valid = false
-		result.Errors = append(result.Errors, fmt.Sprintf("telemetry listen address has invalid format: %v (expected 'host:port', e.g. '0.0.0.0:8090' or '[::1]:8090')", err))
-		return result
-	}
-
-	port, err := strconv.Atoi(portStr)
-	if err != nil {
-		result.Valid = false
-		result.Errors = append(result.Errors, fmt.Sprintf("telemetry listen port is not a valid number: %s", portStr))
-		return result
-	}
-
-	if port < 1 || port > 65535 {
-		result.Valid = false
-		result.Errors = append(result.Errors, fmt.Sprintf("telemetry listen port must be between 1 and 65535, got %d", port))
-	}
-
-	return result
 }
 
 // ValidateSettings validates the entire Settings struct
@@ -843,1175 +155,4 @@ func ValidateSettings(settings *Settings) error {
 		return ve
 	}
 	return nil
-}
-
-// validateBirdNETSettings validates the BirdNET-specific settings.
-// This function uses ValidateBirdNETSettings internally and handles side effects
-// (logging, mutation) to maintain backward compatibility.
-func validateBirdNETSettings(birdnetSettings *BirdNETConfig, settings *Settings) error {
-	// Call the pure validation function
-	result := ValidateBirdNETSettings(birdnetSettings)
-
-	// Apply normalized configuration (side effect: mutation)
-	if normalized, ok := result.Normalized.(*BirdNETConfig); ok && normalized != nil {
-		*birdnetSettings = *normalized
-	} else if !ok {
-		// Type assertion failed - this indicates a bug in ValidateBirdNETSettings
-		return errors.Newf("internal error: ValidateBirdNETSettings returned unexpected type %T", result.Normalized).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "birdnet-type-assertion").
-			Build()
-	}
-
-	// Handle warnings (side effects: logging)
-	// Locale fallback warnings are debug-level since the fallback works correctly
-	// and common locales like "en" or "en-US" always resolve to "en-uk"
-	for _, warning := range result.Warnings {
-		GetLogger().Debug("Configuration notice", logger.String("message", warning))
-	}
-
-	// Return errors if validation failed
-	if !result.Valid {
-		return errors.Newf("birdnet settings errors: %v", result.Errors).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "birdnet-settings-collection").
-			Build()
-	}
-
-	return nil
-}
-
-// validateWebServerSettings validates the WebServer-specific settings.
-// This function uses ValidateWebServerSettings internally and handles error formatting
-// to maintain backward compatibility.
-func validateWebServerSettings(settings *WebServerSettings) error {
-	// Call the pure validation function
-	result := ValidateWebServerSettings(settings)
-
-	// Return errors if validation failed
-	if !result.Valid {
-		// Format first error with enhanced error for backward compatibility
-		firstError := result.Errors[0]
-		return errors.Newf("%s", firstError).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "webserver-settings").
-			Build()
-	}
-
-	return nil
-}
-
-// validateSecuritySettings validates the security-specific settings
-func validateSecuritySettings(settings *Security) error {
-	// Check if any OAuth provider is enabled (OAuth providers require host or baseUrl for redirect URLs)
-	// Note: BasicAuth doesn't require host as it doesn't use OAuth redirects
-	if (settings.GoogleAuth.Enabled || settings.GithubAuth.Enabled || settings.MicrosoftAuth.Enabled) && settings.Host == "" && settings.BaseURL == "" {
-		return errors.Newf("security.host or security.baseUrl must be set when using OAuth authentication providers (Google, GitHub, or Microsoft)").
-			Category(errors.CategoryValidation).
-			Context("validation_type", "security-oauth-host").
-			Context("google_enabled", settings.GoogleAuth.Enabled).
-			Context("github_enabled", settings.GithubAuth.Enabled).
-			Context("microsoft_enabled", settings.MicrosoftAuth.Enabled).
-			Build()
-	}
-
-	// TLS mode validation
-	if err := validateTLSMode(settings); err != nil {
-		return err
-	}
-
-	// Validate the subnet bypass setting against the allowed pattern.
-	// Empty entries (from trailing commas, double commas, or all-whitespace tokens)
-	// are skipped so that a config like "10.0.0.0/8, ,192.168.0.0/24" is accepted
-	// with the same semantics as oauth.go's allowlist check.
-	if settings.AllowSubnetBypass.Enabled {
-		subnets := strings.SplitSeq(settings.AllowSubnetBypass.Subnet, ",")
-		for subnet := range subnets {
-			trimmedSubnet := strings.TrimSpace(subnet)
-			if trimmedSubnet == "" {
-				continue // Skip empty entries (e.g. trailing or embedded commas)
-			}
-			_, _, err := net.ParseCIDR(trimmedSubnet)
-			if err != nil {
-				return errors.New(err).
-					Category(errors.CategoryValidation).
-					Context("validation_type", "security-subnet-format").
-					Context("subnet", trimmedSubnet).
-					Build()
-			}
-		}
-	}
-
-	// Validate session duration
-	if settings.SessionDuration <= 0 {
-		return errors.Newf("security.sessionduration must be a positive duration").
-			Category(errors.CategoryValidation).
-			Context("validation_type", "security-session-duration").
-			Build()
-	}
-
-	// Validate OIDC provider configuration
-	if err := validateOIDCProviders(settings); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// validateTLSMode validates TLS certificate management mode settings.
-func validateTLSMode(settings *Security) error {
-	switch settings.TLSMode {
-	case TLSModeAutoTLS:
-		hostname := settings.GetHostnameForCertificates()
-		if hostname == "" {
-			return errors.Newf("security.host (or hostname in security.baseUrl) must be set when TLS mode is autotls").
-				Category(errors.CategoryValidation).
-				Context("validation_type", "security-autotls-host").
-				Build()
-		}
-		if err := validateAutoTLSHostname(hostname); err != nil {
-			return err
-		}
-		if RunningInContainer() {
-			GetLogger().Warn("AutoTLS requires ports 80 and 443 to be exposed",
-				logger.String("ports", "80:80 (ACME HTTP-01), 443:443 (HTTPS)"),
-				logger.String("hint", "Consider using docker-compose.autotls.yml for proper AutoTLS configuration"))
-		}
-
-	case TLSModeManual:
-		tm := GetTLSManager()
-		if !tm.CertificateExists("webserver", TLSCertTypeServerCert) {
-			GetLogger().Warn("TLS mode is 'manual' but no server certificate is installed",
-				logger.String("hint", "Upload a server certificate via the settings page or API"))
-		}
-
-	case TLSModeSelfSigned:
-		tm := GetTLSManager()
-		if !tm.CertificateExists("webserver", TLSCertTypeServerCert) {
-			GetLogger().Info("TLS mode is 'selfsigned' - a self-signed certificate will be generated on startup")
-		}
-
-	case TLSModeNone:
-		// No TLS validation needed
-
-	default:
-		return errors.Newf("security.tlsMode has invalid value %q (valid: autotls, manual, selfsigned, or empty)", settings.TLSMode).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "security-tlsmode-invalid").
-			Context("tls_mode", string(settings.TLSMode)).
-			Build()
-	}
-	return nil
-}
-
-// validateOIDCProviders validates OIDC-specific provider configuration:
-// at most one OIDC entry, valid issuer URL when enabled, callback URL source required, HTTPS warning.
-func validateOIDCProviders(settings *Security) error {
-	oidcCount := 0
-	for i := range settings.OAuthProviders {
-		provider := &settings.OAuthProviders[i]
-		if provider.Provider != "oidc" {
-			continue
-		}
-		oidcCount++
-		if oidcCount > 1 {
-			return errors.Newf("only one OIDC provider (provider: \"oidc\") is allowed in security.oauthProviders, found duplicate entry").
-				Category(errors.CategoryValidation).
-				Context("validation_type", "security-oidc-duplicate").
-				Build()
-		}
-		if !provider.Enabled {
-			continue
-		}
-		// Require a callback URL source — without host/baseUrl/redirectUri, the OAuth flow will fail
-		if provider.RedirectURI == "" && settings.Host == "" && settings.BaseURL == "" {
-			return errors.Newf("security.host or security.baseUrl must be set, or redirectUri must be configured, when OIDC provider is enabled").
-				Category(errors.CategoryValidation).
-				Context("validation_type", "security-oidc-redirect-missing").
-				Context("issuer_url", provider.IssuerURL).
-				Build()
-		}
-		if provider.IssuerURL == "" {
-			return errors.Newf("security.oauthProviders: issuerUrl is required when provider is \"oidc\" and enabled").
-				Category(errors.CategoryValidation).
-				Context("validation_type", "security-oidc-issuer-missing").
-				Build()
-		}
-		parsed, err := url.Parse(provider.IssuerURL)
-		if err != nil || parsed.Host == "" || (parsed.Scheme != SchemeHTTPS && parsed.Scheme != SchemeHTTP) {
-			return errors.Newf("security.oauthProviders: issuerUrl %q is not a valid URL", provider.IssuerURL).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "security-oidc-issuer-invalid").
-				Context("issuer_url", provider.IssuerURL).
-				Build()
-		}
-		if parsed.Scheme == SchemeHTTP {
-			GetLogger().Warn("OIDC issuerUrl uses HTTP instead of HTTPS — acceptable for local development only",
-				logger.String("issuer_url", provider.IssuerURL))
-		}
-	}
-	return nil
-}
-
-// privateTLDs lists TLD suffixes that are not publicly resolvable
-// and therefore cannot be used with Let's Encrypt.
-var privateTLDs = []string{
-	".local",
-	".internal",
-	".lan",
-	".home",
-	".localdomain",
-	".localhost",
-	".test",
-	".example",
-	".invalid",
-}
-
-// validateAutoTLSHostname checks that a hostname is suitable for Let's Encrypt.
-// Let's Encrypt requires a publicly resolvable FQDN — not an IP, not a private
-// name, and not a bare hostname without dots.
-func validateAutoTLSHostname(hostname string) error {
-	// Must not be an IP address
-	if net.ParseIP(hostname) != nil {
-		return errors.Newf("Let's Encrypt requires a domain name, not an IP address (%s)", hostname).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "security-autotls-hostname").
-			Context("hostname", hostname).
-			Build()
-	}
-
-	// Must not be localhost (check before dot check since "localhost" has no dots)
-	if strings.EqualFold(hostname, "localhost") {
-		return errors.Newf("Let's Encrypt cannot issue certificates for localhost").
-			Category(errors.CategoryValidation).
-			Context("validation_type", "security-autotls-hostname").
-			Context("hostname", hostname).
-			Build()
-	}
-
-	// Must contain at least one dot (FQDN)
-	if !strings.Contains(hostname, ".") {
-		return errors.Newf("Let's Encrypt requires a fully qualified domain name (e.g., birds.example.com), not a bare hostname (%s)", hostname).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "security-autotls-hostname").
-			Context("hostname", hostname).
-			Build()
-	}
-
-	// Must not use a private/non-routable TLD
-	lower := strings.ToLower(hostname)
-	for _, suffix := range privateTLDs {
-		if strings.HasSuffix(lower, suffix) {
-			return errors.Newf("Let's Encrypt cannot issue certificates for private domain %q (TLD %s is not publicly resolvable)", hostname, suffix).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "security-autotls-hostname").
-				Context("hostname", hostname).
-				Context("tld", suffix).
-				Build()
-		}
-	}
-
-	return nil
-}
-
-// validateRealtimeSettings validates the Realtime-specific settings
-func validateRealtimeSettings(settings *RealtimeSettings) error {
-	// Check if interval is non-negative
-	if settings.Interval < 0 {
-		return errors.Newf("realtime interval must be non-negative").
-			Category(errors.CategoryValidation).
-			Context("validation_type", "realtime-interval").
-			Build()
-	}
-
-	// Validate MQTT settings
-	if err := validateMQTTSettings(&settings.MQTT); err != nil {
-		return err
-	}
-
-	// Validate sound level settings
-	if err := validateSoundLevelSettings(&settings.Audio.SoundLevel); err != nil {
-		return err
-	}
-
-	// Validate species settings
-	if err := validateSpeciesConfigSettings(&settings.Species); err != nil {
-		return err
-	}
-
-	// Apply default transport before stream validation
-	settings.RTSP.ApplyStreamDefaults()
-
-	// Validate stream configurations
-	if err := settings.RTSP.ValidateStreams(); err != nil {
-		return errors.New(err).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "stream-config").
-			Build()
-	}
-
-	// Validate telemetry settings
-	if err := validateTelemetrySettings(&settings.Telemetry); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// validateMQTTSettings validates the MQTT-specific settings.
-// This function uses ValidateMQTTSettings internally and handles error formatting
-// to maintain backward compatibility.
-func validateMQTTSettings(settings *MQTTSettings) error {
-	// Call the pure validation function
-	result := ValidateMQTTSettings(settings)
-
-	// Return errors if validation failed
-	if !result.Valid {
-		// Format first error with enhanced error for backward compatibility
-		firstError := result.Errors[0]
-		return errors.Newf("%s", firstError).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "mqtt-settings").
-			Build()
-	}
-
-	return nil
-}
-
-// validateTelemetrySettings validates the telemetry-specific settings.
-// This function uses ValidateTelemetrySettings internally and handles error formatting.
-func validateTelemetrySettings(settings *TelemetrySettings) error {
-	result := ValidateTelemetrySettings(settings)
-	if !result.Valid {
-		return errors.Newf("%s", result.Errors[0]).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "telemetry-listen-address").
-			Build()
-	}
-	return nil
-}
-
-// validateSoundLevelSettings validates the SoundLevel-specific settings
-func validateSoundLevelSettings(settings *SoundLevelSettings) error {
-	// Sound level settings are optional, only validate if enabled
-	if settings.Enabled {
-		// Check if interval is at least the minimum to avoid excessive CPU usage
-		if settings.Interval < MinSoundLevelInterval {
-			return errors.Newf("sound level interval must be at least %d seconds to avoid excessive CPU usage, got %d", MinSoundLevelInterval, settings.Interval).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "sound-level-interval").
-				Context("interval", settings.Interval).
-				Context("minimum_interval", MinSoundLevelInterval).
-				Build()
-		}
-	}
-	return nil
-}
-
-// validateBirdweatherSettings validates the Birdweather-specific settings.
-// This function uses ValidateBirdweatherSettings internally and handles side effects
-// (logging, mutation) to maintain backward compatibility.
-func validateBirdweatherSettings(settings *BirdweatherSettings) error {
-	// Call the pure validation function
-	result := ValidateBirdweatherSettings(settings)
-
-	// Apply normalized configuration (side effect: mutation)
-	if normalized, ok := result.Normalized.(*BirdweatherSettings); ok && normalized != nil {
-		*settings = *normalized
-	} else if !ok {
-		// Type assertion failed - this indicates a bug in ValidateBirdweatherSettings
-		return errors.Newf("internal error: ValidateBirdweatherSettings returned unexpected type %T", result.Normalized).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "birdweather-type-assertion").
-			Build()
-	}
-
-	// Handle warnings (side effect: logging)
-	for _, warning := range result.Warnings {
-		GetLogger().Warn("Birdweather validation warning", logger.String("message", warning))
-	}
-
-	// Return errors if validation failed
-	if !result.Valid {
-		// Join errors for backward compatibility
-		return errors.Newf("%s", strings.Join(result.Errors, "; ")).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "birdweather-settings").
-			Build()
-	}
-
-	return nil
-}
-
-// validateAudioSettings validates the audio settings and sets ffmpeg and sox paths
-func validateAudioSettings(settings *AudioSettings) error {
-	// Validate and determine the effective FFmpeg path
-	validatedFfmpegPath, ffmpegErr := ValidateToolPath(settings.FfmpegPath, GetFfmpegBinaryName())
-	if ffmpegErr != nil {
-		GetLogger().Warn("FFmpeg validation failed", logger.Error(ffmpegErr), logger.String("impact", "Audio export/conversion requiring FFmpeg might be disabled or use defaults"))
-		// Log validation warning for telemetry
-		logValidationWarning(ffmpegErr, "audio-tool-ffmpeg", "ffmpeg-not-available")
-		settings.FfmpegPath = "" // Ensure path is empty if validation failed
-	} else {
-		settings.FfmpegPath = validatedFfmpegPath // Store the validated path (explicit or from PATH)
-
-		// Detect FFmpeg version for runtime decisions (e.g., FFmpeg 5.x bug workarounds)
-		version, major, minor := GetFfmpegVersion()
-		settings.FfmpegVersion = version
-		settings.FfmpegMajor = major
-		settings.FfmpegMinor = minor
-
-		if major > 0 {
-			GetLogger().Debug("Detected FFmpeg version", logger.String("version", version), logger.Int("major", major), logger.Int("minor", minor))
-		} else {
-			GetLogger().Warn("Could not detect FFmpeg version", logger.String("version_string", version))
-		}
-	}
-
-	// Validate and determine the effective SoX path
-	// We only need to know if it's available and its formats, so LookPath is sufficient here.
-	soxPath, soxLookPathErr := exec.LookPath(GetSoxBinaryName())
-	if soxLookPathErr != nil {
-		settings.SoxPath = ""
-		settings.SoxAudioTypes = nil
-		GetLogger().Warn("SoX not found in system PATH", logger.String("impact", "Audio source processing requiring SoX might be disabled"))
-	} else {
-		settings.SoxPath = soxPath
-		// Get supported formats if SoX is found
-		_, formats := IsSoxAvailable() // We already know it's available from LookPath
-		settings.SoxAudioTypes = formats
-	}
-
-	// Validate audio sources
-	if err := settings.ValidateSources(); err != nil {
-		return errors.New(err).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-sources").
-			Build()
-	}
-
-	// Validate global quiet hours (legacy fallback)
-	if err := ValidateQuietHours(&settings.QuietHours, "sound card"); err != nil {
-		return errors.New(err).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-quiet-hours").
-			Build()
-	}
-
-	// Validate audio export settings.
-	//
-	// Normalize first, validate second. Type must be normalized even when
-	// Export.Enabled is false, because the Enabled flag can be flipped back on
-	// without triggering validation of the already-persisted Type value.
-	// An empty Type previously produced extension-less clip_name rows in the
-	// DB, causing audio download and spectrogram generation to 404
-	// (GitHub #2810, #2814).
-	if strings.TrimSpace(settings.Export.Type) == "" {
-		GetLogger().Warn("audio export type is empty, normalizing to wav",
-			logger.String("previous_type", settings.Export.Type),
-			logger.String("action", "normalize_to_wav"))
-		settings.Export.Type = AudioExportTypeWAV
-	}
-
-	// Reject unknown Type before any fallback runs. Doing this first means a
-	// garbage value cannot be silently rewritten to WAV by the ffmpeg-missing
-	// fallback below, which would hide the misconfiguration.
-	switch settings.Export.Type {
-	case AudioExportTypeWAV, AudioExportTypeFLAC,
-		AudioExportTypeAAC, AudioExportTypeOPUS, AudioExportTypeMP3:
-		// known format
-	default:
-		return errors.Newf("unsupported audio export type: %s", settings.Export.Type).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-export-type").
-			Context("export_type", settings.Export.Type).
-			Build()
-	}
-
-	if settings.FfmpegPath == "" && settings.Export.Type != AudioExportTypeWAV {
-		GetLogger().Warn("FFmpeg not available, forcing WAV format for audio export",
-			logger.String("previous_type", settings.Export.Type))
-		settings.Export.Type = AudioExportTypeWAV
-	}
-
-	// Bitrate only matters for lossy formats and only when export is enabled.
-	switch settings.Export.Type {
-	case AudioExportTypeAAC, AudioExportTypeOPUS, AudioExportTypeMP3:
-		if settings.Export.Enabled {
-			if err := validateExportBitrate(settings.Export.Type, settings.Export.Bitrate); err != nil {
-				return err
-			}
-		}
-	}
-
-	// Remaining checks (length, pre-capture, gain, normalization) only make
-	// sense when export is actually enabled.
-	if settings.Export.Enabled {
-		// Validate capture length (10-60 seconds)
-		if settings.Export.Length < 10 || settings.Export.Length > 60 {
-			return errors.Newf("audio capture length must be between 10 and 60 seconds, got %d", settings.Export.Length).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "audio-export-capture-length").
-				Context("capture_length", settings.Export.Length).
-				Build()
-		}
-
-		// Validate pre-capture (max 1/2 of capture length)
-		maxPreCapture := settings.Export.Length / 2
-		if settings.Export.PreCapture < 0 || settings.Export.PreCapture > maxPreCapture {
-			return errors.Newf("audio pre-capture must be between 0 and %d seconds (1/2 of capture length), got %d", maxPreCapture, settings.Export.PreCapture).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "audio-export-precapture").
-				Context("precapture", settings.Export.PreCapture).
-				Context("max_precapture", maxPreCapture).
-				Context("capture_length", settings.Export.Length).
-				Build()
-		}
-
-		// Validate gain setting (reasonable range for audio processing)
-		if settings.Export.Gain < MinAudioGain || settings.Export.Gain > MaxAudioGain {
-			return errors.Newf("audio gain must be between %.0f and +%.0f dB, got %.1f", MinAudioGain, MaxAudioGain, settings.Export.Gain).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "audio-export-gain").
-				Context("gain", settings.Export.Gain).
-				Context("min_gain", MinAudioGain).
-				Context("max_gain", MaxAudioGain).
-				Build()
-		}
-
-		// Validate normalization settings if enabled
-		if settings.Export.Normalization.Enabled {
-			if err := validateNormalizationSettings(&settings.Export.Normalization, settings.Export.Gain); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-// Bitrate constraints for lossy audio export formats (mp3, aac, opus).
-// The suffix and kbps bounds are consumed by validateExportBitrate and
-// documented in the frontend's getBitrateConfig (audioValidation.ts).
-const (
-	audioExportBitrateSuffix  = "k"
-	minAudioExportBitrateKbps = 32
-	maxAudioExportBitrateKbps = 320
-)
-
-// validateExportBitrate validates the bitrate setting for lossy audio export
-// formats. The value must use the "k" suffix and fall within the supported
-// kbps range.
-func validateExportBitrate(exportType, bitrate string) error {
-	if !strings.HasSuffix(bitrate, audioExportBitrateSuffix) {
-		return errors.Newf("invalid bitrate format for %s: %s. Must end with %q (e.g., '64k')", exportType, bitrate, audioExportBitrateSuffix).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-export-bitrate-format").
-			Context("export_type", exportType).
-			Context("bitrate", bitrate).
-			Build()
-	}
-	bitrateValue, err := strconv.Atoi(strings.TrimSuffix(bitrate, audioExportBitrateSuffix))
-	if err != nil {
-		return errors.Newf("invalid bitrate value for %s: %s", exportType, bitrate).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-export-bitrate-value").
-			Context("export_type", exportType).
-			Context("bitrate", bitrate).
-			Build()
-	}
-	if bitrateValue < minAudioExportBitrateKbps || bitrateValue > maxAudioExportBitrateKbps {
-		return errors.Newf("bitrate for %s must be between %dk and %dk, got %dk",
-			exportType, minAudioExportBitrateKbps, maxAudioExportBitrateKbps, bitrateValue).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-export-bitrate-range").
-			Context("export_type", exportType).
-			Context("bitrate_value", bitrateValue).
-			Build()
-	}
-	return nil
-}
-
-// validateNormalizationSettings validates the EBU R128 normalization parameters
-// (target LUFS, loudness range, and true peak) and warns if gain is also configured.
-func validateNormalizationSettings(norm *NormalizationSettings, gain float64) error {
-	if norm.TargetLUFS < MinTargetLUFS || norm.TargetLUFS > MaxTargetLUFS {
-		return errors.Newf("normalization target LUFS must be between %.0f and %.0f, got %.1f", MinTargetLUFS, MaxTargetLUFS, norm.TargetLUFS).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-normalization-target").
-			Context("target_lufs", norm.TargetLUFS).
-			Context("min_target_lufs", MinTargetLUFS).
-			Context("max_target_lufs", MaxTargetLUFS).
-			Build()
-	}
-	if norm.LoudnessRange < MinLoudnessRange || norm.LoudnessRange > MaxLoudnessRange {
-		return errors.Newf("normalization loudness range must be between %.0f and %.0f LU, got %.1f", MinLoudnessRange, MaxLoudnessRange, norm.LoudnessRange).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-normalization-range").
-			Context("loudness_range", norm.LoudnessRange).
-			Context("min_loudness_range", MinLoudnessRange).
-			Context("max_loudness_range", MaxLoudnessRange).
-			Build()
-	}
-	if norm.TruePeak < MinTruePeak || norm.TruePeak > MaxTruePeak {
-		return errors.Newf("normalization true peak must be between %.0f and %.0f dBTP, got %.1f", MinTruePeak, MaxTruePeak, norm.TruePeak).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "audio-normalization-peak").
-			Context("true_peak", norm.TruePeak).
-			Context("min_true_peak", MinTruePeak).
-			Context("max_true_peak", MaxTruePeak).
-			Build()
-	}
-	if gain != 0 {
-		GetLogger().Warn("Both gain and normalization are configured", logger.String("action", "Normalization will take precedence, gain setting will be ignored"))
-	}
-	return nil
-}
-
-// validateRetentionSettings validates retention policy, MaxAge, and MaxUsage at startup.
-// This catches invalid values early instead of failing silently at runtime when the
-// disk manager first attempts cleanup.
-func validateRetentionSettings(settings *RetentionSettings) error {
-	// Empty policy means retention is disabled — treat as "none"
-	if settings.Policy == "" {
-		return nil
-	}
-
-	// Validate policy against known values
-	if !slices.Contains(validRetentionPolicies, settings.Policy) {
-		return errors.Newf("retention policy must be one of %v, got %q", validRetentionPolicies, settings.Policy).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "retention-policy").
-			Context("policy", settings.Policy).
-			Build()
-	}
-
-	// Validate MaxAge when age-based policy is active
-	if settings.Policy == RetentionPolicyAge {
-		if _, err := ParseRetentionPeriod(settings.MaxAge); err != nil {
-			return errors.Newf("retention maxAge %q is invalid: %v", settings.MaxAge, err).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "retention-max-age").
-				Context("max_age", settings.MaxAge).
-				Build()
-		}
-	}
-
-	// Validate MaxUsage when usage-based policy is active
-	if settings.Policy == RetentionPolicyUsage {
-		if _, err := ParsePercentage(settings.MaxUsage, "retention.maxUsage"); err != nil {
-			return errors.Newf("retention maxUsage %q is invalid: %v", settings.MaxUsage, err).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "retention-max-usage").
-				Context("max_usage", settings.MaxUsage).
-				Build()
-		}
-	}
-
-	return nil
-}
-
-func validateDashboardSettings(settings *Dashboard) error {
-	// Validate deprecated root SummaryLimit (only when non-zero, i.e. not yet migrated)
-	if settings.SummaryLimit != 0 && (settings.SummaryLimit < 10 || settings.SummaryLimit > 1000) {
-		GetLogger().Warn("Dashboard SummaryLimit out of range (10-1000), resetting to 10",
-			logger.Int("invalid_value", settings.SummaryLimit))
-		settings.SummaryLimit = 10
-	}
-
-	// Validate layout element configs
-	validElementTypes := []string{"banner", "daily-summary", "currently-hearing", "detections-grid", "live-spectrogram", "video-embed"}
-	validWidths := []string{"", "full", "half"}
-
-	for i, el := range settings.Layout.Elements {
-		if !slices.Contains(validElementTypes, el.Type) {
-			return errors.Newf("Dashboard layout element %d has invalid type %q", i, el.Type).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "dashboard-element-type").
-				Context("element_index", i).
-				Context("element_type", el.Type).
-				Build()
-		}
-
-		if !slices.Contains(validWidths, el.Width) {
-			return errors.Newf("Dashboard layout element %d has invalid width %q (must be \"full\" or \"half\")", i, el.Width).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "dashboard-element-width").
-				Context("element_index", i).
-				Context("element_width", el.Width).
-				Build()
-		}
-
-		if el.Summary != nil && el.Summary.SummaryLimit != 0 && (el.Summary.SummaryLimit < 10 || el.Summary.SummaryLimit > 1000) {
-			GetLogger().Warn("Dashboard layout element SummaryLimit out of range (10-1000), resetting to 10",
-				logger.Int("element_index", i),
-				logger.String("element_type", el.Type),
-				logger.Int("invalid_value", el.Summary.SummaryLimit))
-			el.Summary.SummaryLimit = 10
-			settings.Layout.Elements[i] = el
-		}
-	}
-
-	// Validate UI locale if provided
-	if settings.Locale != "" {
-		isValid := slices.Contains(ValidUILocales(), settings.Locale)
-		if !isValid {
-			// Log warning but don't fail - fallback to default
-			GetLogger().Warn("Invalid UI locale, will use default", logger.String("invalid_locale", settings.Locale), logger.String("fallback", "en"))
-			settings.Locale = "en"
-		}
-	}
-
-	// Validate default audio gain (playback UI preference)
-	if settings.DefaultAudioGain < MinPlaybackGain || settings.DefaultAudioGain > MaxPlaybackGain {
-		GetLogger().Warn("Dashboard DefaultAudioGain out of range, clamping to safe value",
-			logger.Float64("invalid_value", settings.DefaultAudioGain),
-			logger.Float64("min", MinPlaybackGain),
-			logger.Float64("max", MaxPlaybackGain))
-		if settings.DefaultAudioGain < MinPlaybackGain {
-			settings.DefaultAudioGain = MinPlaybackGain
-		} else {
-			settings.DefaultAudioGain = MaxPlaybackGain
-		}
-	}
-
-	// Validate spectrogram settings
-	if settings.Spectrogram.Mode != "" {
-		validModes := []string{"auto", "prerender", "user-requested"}
-		isValid := slices.Contains(validModes, settings.Spectrogram.Mode)
-		if !isValid {
-			// Log warning but don't fail - GetMode() will handle fallback
-			GetLogger().Warn("Invalid spectrogram mode, using GetMode() fallback",
-				logger.String("invalid_mode", settings.Spectrogram.Mode),
-				logger.String("valid_modes", "auto, prerender, user-requested"))
-		}
-	}
-
-	// Validate spectrogram size
-	if settings.Spectrogram.Size != "" {
-		validSizes := []string{"sm", "md", "lg", "xl"}
-		if !slices.Contains(validSizes, settings.Spectrogram.Size) {
-			GetLogger().Warn("Invalid spectrogram size, using default",
-				logger.String("invalid_size", settings.Spectrogram.Size),
-				logger.String("valid_sizes", strings.Join(validSizes, ", ")),
-				logger.String("fallback", "lg"))
-			settings.Spectrogram.Size = "lg"
-		}
-	}
-
-	// Validate spectrogram style
-	if settings.Spectrogram.Style != "" {
-		validStyles := []string{
-			SpectrogramStyleDefault,
-			SpectrogramStyleScientificDark,
-			SpectrogramStyleHighContrastDark,
-			SpectrogramStyleScientific,
-		}
-		if !slices.Contains(validStyles, settings.Spectrogram.Style) {
-			// Log warning but don't fail - default to "default" style
-			GetLogger().Warn("Invalid spectrogram style, using default",
-				logger.String("invalid_style", settings.Spectrogram.Style),
-				logger.String("valid_styles", strings.Join(validStyles, ", ")))
-			settings.Spectrogram.Style = SpectrogramStyleDefault
-		}
-	}
-
-	// Log the effective spectrogram mode at startup for troubleshooting
-	effectiveMode := settings.Spectrogram.GetMode()
-	GetLogger().Debug("Spectrogram configuration",
-		logger.Bool("enabled", settings.Spectrogram.Enabled),
-		logger.String("mode", settings.Spectrogram.Mode),
-		logger.String("effective_mode", effectiveMode),
-		logger.String("size", settings.Spectrogram.Size),
-		logger.Bool("raw", settings.Spectrogram.Raw),
-		logger.String("style", settings.Spectrogram.Style))
-
-	return nil
-}
-
-// validateWeatherSettings validates weather-specific settings
-func validateWeatherSettings(settings *WeatherSettings) error {
-	// Validate poll interval (minimum 15 minutes)
-	if settings.PollInterval < 15 {
-		return errors.Newf("weather poll interval must be at least 15 minutes, got %d", settings.PollInterval).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "weather-poll-interval").
-			Context("poll_interval", settings.PollInterval).
-			Build()
-	}
-
-	// Validate Wunderground settings if it's the selected provider
-	if settings.Provider == "wunderground" {
-		if err := settings.Wunderground.ValidateWunderground(); err != nil {
-			return errors.New(err).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "wunderground-settings").
-				Build()
-		}
-	}
-
-	return nil
-}
-
-// validateSpeciesTrackingSettings validates the species tracking settings
-func validateSpeciesTrackingSettings(settings *SpeciesTrackingSettings) error {
-	if settings.Enabled {
-		// Validate window days
-		if settings.NewSpeciesWindowDays < 1 || settings.NewSpeciesWindowDays > 365 {
-			return errors.Newf("species tracking window days must be between 1 and 365, got %d", settings.NewSpeciesWindowDays).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "species-tracking-window-days").
-				Context("window_days", settings.NewSpeciesWindowDays).
-				Build()
-		}
-
-		// Validate sync interval
-		if settings.SyncIntervalMinutes < 5 || settings.SyncIntervalMinutes > 1440 {
-			return errors.Newf("species tracking sync interval must be between 5 and 1440 minutes (24 hours), got %d", settings.SyncIntervalMinutes).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "species-tracking-sync-interval").
-				Context("sync_interval", settings.SyncIntervalMinutes).
-				Build()
-		}
-
-		// Validate notification suppression hours
-		if settings.NotificationSuppressionHours < 0 || settings.NotificationSuppressionHours > 720 {
-			return errors.Newf("notification suppression hours must be between 0 and 720 (30 days), got %d", settings.NotificationSuppressionHours).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "notification-suppression-hours").
-				Context("suppression_hours", settings.NotificationSuppressionHours).
-				Build()
-		}
-
-		// Validate yearly tracking settings
-		if err := validateYearlyTrackingSettings(&settings.YearlyTracking); err != nil {
-			return err
-		}
-
-		// Validate seasonal tracking settings
-		if err := validateSeasonalTrackingSettings(&settings.SeasonalTracking); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateYearlyTrackingSettings(settings *YearlyTrackingSettings) error {
-	if settings.Enabled {
-		// Validate reset month
-		if settings.ResetMonth < 1 || settings.ResetMonth > 12 {
-			return errors.Newf("yearly tracking reset month must be between 1 and 12, got %d", settings.ResetMonth).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "yearly-tracking-reset-month").
-				Context("reset_month", settings.ResetMonth).
-				Build()
-		}
-		// Validate reset day - must be valid for the specified month
-		maxDaysInMonth := getMaxDaysInMonth(settings.ResetMonth)
-		if settings.ResetDay < 1 || settings.ResetDay > maxDaysInMonth {
-			return errors.Newf("yearly tracking reset day must be between 1 and %d for month %d, got %d", maxDaysInMonth, settings.ResetMonth, settings.ResetDay).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "yearly-tracking-reset-day").
-				Context("reset_month", settings.ResetMonth).
-				Context("reset_day", settings.ResetDay).
-				Context("max_days_in_month", maxDaysInMonth).
-				Build()
-		}
-		// Validate window days
-		if settings.WindowDays < 1 || settings.WindowDays > 365 {
-			return errors.Newf("yearly tracking window days must be between 1 and 365, got %d", settings.WindowDays).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "yearly-tracking-window-days").
-				Context("window_days", settings.WindowDays).
-				Build()
-		}
-	}
-	return nil
-}
-
-func validateSeasonalTrackingSettings(settings *SeasonalTrackingSettings) error {
-	if settings.Enabled {
-		// Validate window days
-		if settings.WindowDays < 1 || settings.WindowDays > 365 {
-			return errors.Newf("seasonal tracking window days must be between 1 and 365, got %d", settings.WindowDays).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "seasonal-tracking-window-days").
-				Context("window_days", settings.WindowDays).
-				Build()
-		}
-		// Validate seasons
-		if len(settings.Seasons) == 0 {
-			return errors.Newf("seasonal tracking requires at least one season to be defined").
-				Category(errors.CategoryValidation).
-				Context("validation_type", "seasonal-tracking-seasons").
-				Build()
-		}
-		for seasonName, season := range settings.Seasons {
-			if season.StartMonth < 1 || season.StartMonth > 12 {
-				return errors.Newf("season %s start month must be between 1 and 12, got %d", seasonName, season.StartMonth).
-					Category(errors.CategoryValidation).
-					Context("validation_type", "seasonal-tracking-season-month").
-					Context("season", seasonName).
-					Context("start_month", season.StartMonth).
-					Build()
-			}
-			maxDaysInMonth := getMaxDaysInMonth(season.StartMonth)
-			if season.StartDay < 1 || season.StartDay > maxDaysInMonth {
-				return errors.Newf("season %s start day must be between 1 and %d for month %d, got %d", seasonName, maxDaysInMonth, season.StartMonth, season.StartDay).
-					Category(errors.CategoryValidation).
-					Context("validation_type", "seasonal-tracking-season-day").
-					Context("season", seasonName).
-					Context("start_month", season.StartMonth).
-					Context("start_day", season.StartDay).
-					Context("max_days_in_month", maxDaysInMonth).
-					Build()
-			}
-		}
-	}
-	return nil
-}
-
-// getMaxDaysInMonth returns the maximum number of days for a given month (1-12)
-func getMaxDaysInMonth(month int) int {
-	switch month {
-	case 2: // February
-		return 29 // Return 29 to safely accommodate leap years, ensuring validation doesn't reject valid Feb 29 dates
-	case 4, 6, 9, 11: // April, June, September, November
-		return 30
-	default: // January, March, May, July, August, October, December
-		return 31
-	}
-}
-
-// validateSpeciesConfigSettings validates the species-specific configuration settings
-func validateSpeciesConfigSettings(settings *SpeciesSettings) error {
-	// Validate each species configuration
-	for speciesName, config := range settings.Config {
-		// Check if interval is non-negative
-		if config.Interval < 0 {
-			return errors.Newf("species config for '%s': interval must be non-negative, got %d", speciesName, config.Interval).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "species-config-interval").
-				Context("species_name", speciesName).
-				Context("interval", config.Interval).
-				Build()
-		}
-
-		// Check if threshold is within valid range
-		if config.Threshold < 0 || config.Threshold > 1 {
-			return errors.Newf("species config for '%s': threshold must be between 0 and 1, got %f", speciesName, config.Threshold).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "species-config-threshold").
-				Context("species_name", speciesName).
-				Context("threshold", config.Threshold).
-				Build()
-		}
-	}
-	return nil
-}
-
-// validateNotificationSettings validates notification push configuration
-func validateNotificationSettings(n *NotificationConfig) error {
-	if !n.Push.Enabled {
-		return nil
-	}
-	// Basic sanity checks
-	if n.Push.MaxRetries < 0 {
-		return errors.Newf("notification.push.max_retries must be >= 0").
-			Category(errors.CategoryValidation).
-			Context("validation_type", "notification-push-max-retries").
-			Build()
-	}
-	if n.Push.DefaultTimeout < 0 || n.Push.RetryDelay < 0 {
-		return errors.Newf("notification push durations must be non-negative").
-			Category(errors.CategoryValidation).
-			Context("validation_type", "notification-push-durations").
-			Build()
-	}
-	for i := range n.Push.Providers {
-		p := &n.Push.Providers[i]
-		ptype := strings.ToLower(p.Type)
-		switch ptype {
-		case "script":
-			if p.Enabled && strings.TrimSpace(p.Command) == "" {
-				return errors.Newf("script provider requires command when enabled").
-					Category(errors.CategoryValidation).
-					Context("validation_type", "notification-push-script-command").
-					Build()
-			}
-		case "shoutrrr":
-			if p.Enabled && len(p.URLs) == 0 {
-				return errors.Newf("shoutrrr provider requires at least one URL when enabled").
-					Category(errors.CategoryValidation).
-					Context("validation_type", "notification-push-shoutrrr-urls").
-					Build()
-			}
-			// Normalize ntfy URLs: ntfy://topic -> ntfy://ntfy.sh/topic
-			normalizeNtfyURLs(p)
-		case "webhook":
-			if err := validateWebhookProvider(p); err != nil {
-				return err
-			}
-		default:
-			return errors.Newf("unknown push provider type: %s", p.Type).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "notification-push-provider-type").
-				Build()
-		}
-	}
-	return nil
-}
-
-// validateWebhookProvider validates webhook provider configuration.
-// This function uses ValidateWebhookProvider internally and handles error formatting
-// to maintain backward compatibility. Authentication validation is still performed separately
-// via validateWebhookAuth due to its complexity.
-func validateWebhookProvider(p *PushProviderConfig) error {
-	// Call the pure validation function
-	result := ValidateWebhookProvider(p)
-
-	// Return early if basic validation failed
-	if !result.Valid {
-		// Format first error with enhanced error for backward compatibility
-		firstError := result.Errors[0]
-		return errors.Newf("%s", firstError).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "notification-push-webhook").
-			Context("provider_name", p.Name).
-			Build()
-	}
-
-	// If disabled or no endpoints, no need to validate auth
-	if !p.Enabled || len(p.Endpoints) == 0 {
-		return nil
-	}
-
-	// Validate authentication for each endpoint
-	// Note: Auth validation is kept separate as it's complex and not yet in pure version
-	for i := range p.Endpoints {
-		endpoint := &p.Endpoints[i]
-		if err := validateWebhookAuth(&endpoint.Auth, p.Name, i); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// validateWebhookAuth validates webhook authentication configuration.
-// Checks that required fields are provided but does NOT resolve secrets here.
-// Secret resolution happens at runtime in the webhook provider.
-func validateWebhookAuth(auth *WebhookAuthConfig, providerName string, endpointIndex int) error {
-	authType := strings.ToLower(auth.Type)
-
-	// Empty auth type defaults to "none" - this is valid
-	if authType == "" || authType == "none" {
-		return nil
-	}
-
-	switch authType {
-	case "bearer":
-		// At least one of token or token_file must be provided
-		if strings.TrimSpace(auth.Token) == "" && strings.TrimSpace(auth.TokenFile) == "" {
-			return errors.Newf("webhook provider '%s' endpoint %d: bearer auth requires token or token_file", providerName, endpointIndex).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "notification-push-webhook-auth-bearer").
-				Context("provider_name", providerName).
-				Context("endpoint_index", endpointIndex).
-				Build()
-		}
-	case "basic":
-		// Check user/user_file
-		if strings.TrimSpace(auth.User) == "" && strings.TrimSpace(auth.UserFile) == "" {
-			return errors.Newf("webhook provider '%s' endpoint %d: basic auth requires user or user_file", providerName, endpointIndex).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "notification-push-webhook-auth-basic").
-				Context("provider_name", providerName).
-				Context("endpoint_index", endpointIndex).
-				Build()
-		}
-		// Check pass/pass_file
-		if strings.TrimSpace(auth.Pass) == "" && strings.TrimSpace(auth.PassFile) == "" {
-			return errors.Newf("webhook provider '%s' endpoint %d: basic auth requires pass or pass_file", providerName, endpointIndex).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "notification-push-webhook-auth-basic").
-				Context("provider_name", providerName).
-				Context("endpoint_index", endpointIndex).
-				Build()
-		}
-	case "custom":
-		// Header name is always required (no file variant)
-		if strings.TrimSpace(auth.Header) == "" {
-			return errors.Newf("webhook provider '%s' endpoint %d: custom auth requires header", providerName, endpointIndex).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "notification-push-webhook-auth-custom").
-				Context("provider_name", providerName).
-				Context("endpoint_index", endpointIndex).
-				Build()
-		}
-		// Check value/value_file
-		if strings.TrimSpace(auth.Value) == "" && strings.TrimSpace(auth.ValueFile) == "" {
-			return errors.Newf("webhook provider '%s' endpoint %d: custom auth requires value or value_file", providerName, endpointIndex).
-				Category(errors.CategoryValidation).
-				Context("validation_type", "notification-push-webhook-auth-custom").
-				Context("provider_name", providerName).
-				Context("endpoint_index", endpointIndex).
-				Build()
-		}
-	default:
-		return errors.Newf("webhook provider '%s' endpoint %d: unsupported auth type: %s", providerName, endpointIndex, authType).
-			Category(errors.CategoryValidation).
-			Context("validation_type", "notification-push-webhook-auth-type").
-			Context("provider_name", providerName).
-			Context("endpoint_index", endpointIndex).
-			Context("auth_type", authType).
-			Build()
-	}
-
-	return nil
-}
-
-// NormalizeNtfyURL fixes a bare ntfy topic URL (ntfy://topic) by inserting
-// the default ntfy.sh host, producing ntfy://ntfy.sh/topic. The shoutrrr
-// library interprets ntfy://topic as hostname="topic" with an empty path,
-// which fails to deliver. URLs that already contain a recognizable host
-// (containing a dot, a colon port, "localhost", or an IP address) are
-// returned unchanged. Non-ntfy URLs are returned as-is.
-func NormalizeNtfyURL(raw string) string {
-	u, err := url.Parse(raw)
-	if err != nil || u.Scheme != "ntfy" {
-		return raw
-	}
-
-	// When the URL already has a path (e.g. ntfy://host/topic), the host
-	// part is unambiguous -- leave it alone.
-	if u.Path != "" {
-		return raw
-	}
-
-	// No path means the URL is either a bare topic (ntfy://mytopic) or a
-	// host with an empty topic (ntfy://localhost). Apply a heuristic: if
-	// the host portion looks like a real hostname or IP, leave it as-is.
-	host := u.Hostname()
-	if strings.Contains(host, ".") || host == "localhost" || net.ParseIP(host) != nil {
-		return raw
-	}
-
-	// Also leave URLs with an explicit port alone (e.g. ntfy://myhost:8080).
-	if u.Port() != "" {
-		return raw
-	}
-
-	// It is a bare topic. Reconstruct the URL with ntfy.sh as the host.
-	u.Path = "/" + u.Host
-	u.Host = "ntfy.sh"
-	return u.String()
-}
-
-// normalizeNtfyURLs repairs bare ntfy topic URLs in a shoutrrr provider's
-// URL list so that shoutrrr receives the expected host/topic format.
-func normalizeNtfyURLs(p *PushProviderConfig) {
-	for i, u := range p.URLs {
-		p.URLs[i] = NormalizeNtfyURL(u)
-	}
 }
